Remove the function of readding frequency to a new file.
